spring - corrigido erros de inicialização e reestruturação de diretórios

+++ src/main/java/com/phillips/saper/bancoquestoes/models/DisciplineModel.java
@@ -13,7 +13,7 @@ public class DisciplineModel {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int idDiscipline;
+    private Long idDiscipline;
 
     private String name;
     private String description;
