+++ src/main/java/com/phillips/saper/bancoquestoes/repositories/MaterialRepository.java
@@ -1,10 +1,11 @@
 package com.phillips.saper.bancoquestoes.repositories;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import com.phillips.saper.bancoquestoes.models.MaterialModel;
 
 @Repository
-public interface MaterialRepository extends JpaRepository<MaterialModel, Integer> {
-    
+public interface MaterialRepository extends JpaRepository<MaterialModel, Long> {
+
 }
