+++ src/main/java/com/phillips/saper/bancoquestoes/repositories/SubjectRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.phillips.saper.bancoquestoes.models.SubjectModel;
 
 @Repository
-public interface SubjectRepository extends JpaRepository<SubjectModel, Integer> {
-    
+public interface SubjectRepository extends JpaRepository<SubjectModel, Long> {
+
 }
