+++ src/main/java/com/phillips/saper/bancoquestoes/repositories/TeacherRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.phillips.saper.bancoquestoes.models.TeacherModel;
 
 @Repository
-public interface TeacherRepository extends JpaRepository<TeacherModel, Integer> {
-    
+public interface TeacherRepository extends JpaRepository<TeacherModel, Long> {
+
 }
