+++ src/main/java/com/phillips/saper/bancoquestoes/repositories/TestRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.phillips.saper.bancoquestoes.models.TestModel;
 
 @Repository
-public interface TestRepository extends JpaRepository<TestModel, Integer> {
-    
+public interface TestRepository extends JpaRepository<TestModel, Long> {
+
 }
